move db helper typedefs into db file
