+++ src/discord-cluster-manager/cogs/admin_cog.py
@@ -13,12 +13,11 @@ import yaml
 from consts import GitHubGPU, ModalGPU
 from discord import app_commands
 from discord.ext import commands, tasks
+from leaderboard_db import LeaderboardItem, SubmissionItem
 from task import LeaderboardTask, make_task
 from ui.misc import ConfirmationView, DeleteConfirmationModal, GPUSelectionView
 from utils import (
     KernelBotError,
-    LeaderboardItem,
-    SubmissionItem,
     leaderboard_name_autocomplete,
     send_discord_message,
     setup_logging,
