+++ src/discord-cluster-manager/cogs/verify_run_cog.py
@@ -14,9 +14,10 @@ from consts import GPU, SubmissionMode, get_gpu_by_name
 from discord import app_commands
 from discord.app_commands import Choice
 from discord.ext import commands
+from leaderboard_db import RunItem, SubmissionItem
 from report import MultiProgressReporter
 from task import make_task
-from utils import RunItem, SubmissionItem, send_discord_message, setup_logging, with_error_handling
+from utils import send_discord_message, setup_logging, with_error_handling
 
 logger = setup_logging()
